repobrain/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "repobrain"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "networkx>=3.2",
    "sentence-transformers>=2.7.0",
    "chromadb>=0.5.0",
    "graphviz>=0.20",
    "langdetect>=1.0.9",
    "click>=8.1",
    "openai>=1.30",
    "pyyaml>=6.0",
    "requests>=2.31",
    "streamlit>=1.35",
    "streamlit-agraph>=0.0.45",
]

[project.scripts]
repobrain = "repobrain.cli.main:cli"

[tool.setuptools.packages.find]
where = [".."]
include = ["repobrain*"]
